+++ CqlSharp/Network/Cluster.cs
@@ -13,6 +13,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using CqlSharp.Logging;
+using CqlSharp.Network.Partition;
+using CqlSharp.Protocol;
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
@@ -20,9 +23,6 @@ using System.Linq;
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
-using CqlSharp.Logging;
-using CqlSharp.Network.Partition;
-using CqlSharp.Protocol;
 
 namespace CqlSharp.Network
 {
@@ -226,7 +226,7 @@ namespace CqlSharp.Network
 
             //setup throttle
             int concurrent = _config.MaxConcurrentQueries <= 0
-                                 ? _nodes.Count*_config.MaxConnectionsPerNode*256
+                                 ? _nodes.Count * _config.MaxConnectionsPerNode * 256
                                  : _config.MaxConcurrentQueries;
 
             logger.LogInfo("Cluster is configured to allow {0} parallel queries", concurrent);
@@ -256,6 +256,10 @@ namespace CqlSharp.Network
                     Connection connection;
                     using (logger.ThreadBinding())
                         connection = _connectionStrategy.GetOrCreateConnection(ConnectionScope.Infrastructure, null);
+
+                    //check if we really got a connection
+                    if (connection == null)
+                        throw new CqlException("Can not obtain connection for maintenance channel");
 
                     //setup event handlers
                     connection.OnConnectionChange += (src, ev) => SetupMaintenanceConnection(logger);
@@ -329,7 +333,7 @@ namespace CqlSharp.Network
                 _name, _release, _cqlVersion, partitioner);
 
             //create list of nodes that make up the cluster, and add the seed
-            var found = new List<Node> {seed};
+            var found = new List<Node> { seed };
 
             //get the peers
             using (
@@ -342,11 +346,11 @@ namespace CqlSharp.Network
                 while (await result.ReadAsync().ConfigureAwait(false))
                 {
                     //create a new node
-                    var newNode = new Node((IPAddress) result["rpc_address"], this)
+                    var newNode = new Node((IPAddress)result["rpc_address"], this)
                                       {
-                                          DataCenter = (string) result["data_center"],
-                                          Rack = (string) result["rack"],
-                                          Tokens = (ISet<string>) result["tokens"]
+                                          DataCenter = (string)result["data_center"],
+                                          Rack = (string)result["rack"],
+                                          Tokens = (ISet<string>)result["tokens"]
                                       };
 
                     //add it if it is in scope
@@ -377,7 +381,7 @@ namespace CqlSharp.Network
 
             var query = new QueryFrame(cql, CqlConsistency.One, null);
             var result =
-                (ResultFrame) await connection.SendRequestAsync(query, logger, 1, false, token).ConfigureAwait(false);
+                (ResultFrame)await connection.SendRequestAsync(query, logger, 1, false, token).ConfigureAwait(false);
             var reader = new CqlDataReader(null, result, null);
 
             logger.LogVerbose("Query {0} returned {1} results", cql, reader.Count);
@@ -424,7 +428,7 @@ namespace CqlSharp.Network
             if (args.Change.Equals(ClusterChange.New))
             {
                 //get the connection from which we received the event
-                var connection = (Connection) source;
+                var connection = (Connection)source;
 
                 //get the new peer
                 using (
@@ -436,11 +440,11 @@ namespace CqlSharp.Network
                 {
                     if (await result.ReadAsync().ConfigureAwait(false))
                     {
-                        var newNode = new Node((IPAddress) result["rpc_address"], this)
+                        var newNode = new Node((IPAddress)result["rpc_address"], this)
                                           {
-                                              DataCenter = (string) result["data_center"],
-                                              Rack = (string) result["rack"],
-                                              Tokens = (ISet<string>) result["tokens"]
+                                              DataCenter = (string)result["data_center"],
+                                              Rack = (string)result["rack"],
+                                              Tokens = (ISet<string>)result["tokens"]
                                           };